:pushpin: Add gomod lockfile and move some logic to its own package
