+++ go.mod
@@ -1,0 +1,5 @@
+module github.com/Omar-V2/game-of-life
+
+go 1.15
+
+require github.com/hajimehoshi/ebiten v1.11.9
